run_mcts(): Set reward of root node to zero
Since the reward refers to the transition from the parent node to this node and the root
has no parent, the reward must always be zero.
In Node's __init__() it is set to None to indicate that the Node is not yet expanded so I
don't want to change the value there.