Compute & record target value estimates from tree search
This is to be able to train the value head properly.
The complexity around the Node class is needed because we also need
value estimates for opponent moves, not just when it is the agent's
turn. Since the tree search is only conducted when the agent is at turn,
we hold on to the search tree by keeping a reference to the root node
(mcts_nodes dict, one tree for each rl agent).
Upon the next move of any player, we select the relevant subtree based
on the conducted action (get_action_subtree()), so we can get a value
estimate from this new tree root. If the original tree search did not
explore that action suifficiently, we run the tree search again to
ensure a good estimate quality (ensure_visit_count()).