ReplayBuffer.add_trajectory(): Calculate value targets
The advantages become obvious considering a multiprocessing environment:
All TrajectoryStates needed for calculating a discounted value target
are already available and must not be transferred to another worker.
This is especially important with large n-step returns that require many
future states. The computation time should be neglectible if the
ReplayBuffer runs in a separate worker since it can happen concurrently.