TrajectoryState: distinguish different belief types
This is important for imperfect information games since the simulation
in the tree search and the training afterwards has potentially more
information: It knows the moves of other players while the normal
playout process only knows observable informations and updates its
beliefs only when the agent is at turn.
Because the training batch can start at any player's turn it is
necessary to record the "dynamics" beliefs separately.