Improve TrajectoryState by using attrs package
This makes it possible to force keyword parameter usage at
instantiation.
Also it enables to use an elegant way to apparently change one value of
a TrajectoryState by copying an existing one, using attrs.evolve().
This makes the code around the value target calculation more robust
since we can get rid of the tuple unpacking which depends on the order
of fields.