Parse hydra configuration into the global config C
This reads the configuration into the global config which is imported in
most other files under the name C.
Where functions are expected, the specified strings are resolved to
actual python function.
Similarly, the networks are instantiated and verified that they are
proper subclasses of these in networks/bases.py.