Improve default action selection function
Always choosing an action greedily is bad for exploration.
The new function draws actions based on the visit count distribution,
processed with softmax. The softmax temperature can be configured to
change with increasing move number.
The configuration options are grouped into a new namespace.
For example, the temperature may drop to zero further in the game,
making the action selection greedy, like in the MuZero chess
implementation.