Change name of config items that load functions
The names were originally designed after how I would name a function in
the code.
The _fn suffix makes it obvious that a function is behind these config
items, in the config as well as in the code. This solution provides the
same names between config and code and seems like a good tradeoff
between code and config readability.