Fill global config C in-place
The in-place filling is required so that all modules can do:
from config import C
Because every import of C should refer to the same object, C must be an
instance initialized at module load time and cannot be assigned
afterwards.

The import convenience is achieved by a slightly akward way of
populating the object after initialisation, because this is a frozen
attrs class, which is not intended to be modified.