Add python venv executable path to mypy config
I am somewhat hesitant to making this change, since it bakes the
assumption where the interpreter is located in the repo.
The change is necessary since I want to include mypy in the pre-commit hooks:
pre-commit runs mypy from an isolated virtual environment that is
missing all the dependencies the project has. It would be possible to
install them in pre-commit's venv too with the flag
additional_dependencies but I really don't want to add another big pytorch
installation just for typechecking.

This decision is only really relevant to other people who want to hack
on the code and run the pre-commit hooks:
They have to make sure venv/bin/python points to a python interpreter
which has all the required dependencies.