Remove beliefs
The beliefs were a not fully-thought out way of transporting information
from previous states into the representation network.
It is better to give the representation network access to the previous
latent:
The latents were involved in a consistency loss, while the beliefs were
not. I fear this could lead to a collapse of the latent while
observation information get pushed into the beliefs.

The remaining references to beliefs across the codebase will be removed
in a bigger rewrite that these parts need anyways.