Change network in/outputs, add beliefs
Beliefs are dead, long live beliefs!

Beliefs are a good idea since they separate the internal game state
representation into a observable (latent) and a hidden (belief) section.

The mistake I made previously was to add the beliefs to the
representation network, with the idea being it could "rectify" wrong
beliefs. The correct approach is that the representation network just
encodes the latent, thus sidestepping the problems outlined in the
commit that removed the beliefs. To combine beliefs from possible
intermediate trajectories, one must search the simulated game tree for
latents identical to the new representation and extract the
corresponding beliefs.

The beliefs are typed Optional since they can be disabled. The
alternative would be using empty Tensors, but this leads to problems
with the tensorboard graph display:
https://github.com/pytorch/pytorch/issues/88707

The Prediction network is changed to output a onehot encoding of the
current player instead of the player type, so the teams can be changed
after training.
Also, the observations may include the current player, information about
teams or how many players are in the current game match. In some games
these information make a difference.