Reimplent NetworkBase.si() correctly
Typechecking revealed some issues, and I think this way of writing it is
cleaner.
The return type still needs to be ignored since mypy insists I have to
return type 'R' - which is not known since it is a TypeVar. I expected
the bound on the TypeVar to resolve this, so maybe this is a mypy bug
that will be fixed someday...