Add current_player and beliefs to Node class
Updating to the new network in/outputs.

The ObservationNode is still broken, I need to think about that a bit
first.

I decided to use an type ignore comment on the argmax tensor instead of
the cast because in the future torch might update its type hints so that
argmax.item() always returns an int. The ignore can then be removed.