Remove ObservationNode
I am not sure at this point how imperfect information belief search will
be implemented. I will probably figure out how to do thatit when the fn/
functions are finished, I am confident that the current architecture is
flexible enough to implement it.
The focus of this project are perfect information multiplayer games
anyways.
My main issue is that the root node is expected to switch between
different player actions, but the idea behind the ObservationNode /
belief search was to switch between different beliefs, which breaks this
assumption.