Remove softmax from from_visit_count() policy fn
The softmax interprets the visit counts as log-probabilities, which is bad:
Repeatedly applying softmax to its result converges either to a flat
distribution with all probabilities being equal or a single value of 1
and zero everywhere else. The exact outcome depends on the softmax
temperature. This happens in MuZero when the selection fn leads to a
visit count distribution roughly resembling the policy and the visit
counts are used to create the target policy.

The correct behavior is to rescale the visit counts into a probability
distribution, which is also idempotent and does not cause the
distribution to converge.
This method also does not need to special case unexpanded nodes.

The normalisation option in the softmax function is removed too, since
its sole purpose was to make the softmax invariant to rescaling of the
visit counts, such as when the total number of visit counts is adjusted.