Add OmegaConf resolver to load python functions
The old way of resolving the methods from the fully qualified path
inside populate_config() was not flexible enough:
The resolved paths were hardcoded and not available to other config
locations. An example is the DrawFromPolicyFn class in fn/action.py,
which needs a policy function upon instantiation. The old way supplied
only the str path to the class instead of the actual python method.