Add TrainingData.dummy for padding short batch trajectories
If the sampled trajectory ends inside the training batch, there must be
some data to fill the remaining trajectory steps: Each batch step is
processed at once in the training code because that is how batching
works.
The invalid dummy data is masked out from the loss calculation by the
is_data attribute.