Rewrite ReplayBuffer 1/3: __init__, __len__, documentation
The new implementation concatenates all trajectory states in a
contiguous buffer. This way the replay buffer size now specifies the
maximum number of trajectory states, instead the number of
trajectories with varying length.
That is more in line with the general understanding in reinforcement
learning which measures the amount of training data in number of
environment (=game) steps.

Also, the deque was replaced with a ring buffer. A deque is implemented
with a linked list and does not provide a good random access time during
sampling.

Both changes enable vast sampling speed improvements, but more on that
in a later commit message:
"Rewrite ReplayBuffer 3/3: sample()"

A tensor cache is added because the trajectories are expected to contain
many tensors of the same data.