Add classmethod to stack TrainingData in the batch dim
When a training batch is sampled from the replay buffer, a number of
trajectory sections are selected. The trajectory steps from these
individual sections must then be stacked in the batch dimension.
This method implements this stacking and reduces boilerplate in the
replay buffer implementation.

The returned TrainingData instance contains a single step of the
training batch, with all states from the individual trajectory sections
stacked together. The complete batch is made up of a sequence of these
instances.