Rewrite ReplayBuffer 3/3: sample()
The change to a contiguous ring buffer makes a rewrite of the sample()
method necessary.
To sample a trajectory subsection from the contiguous buffer, a random
index is drawn in the range of the buffer length. The index is accepted
after verifying that the remaining trajectory starting at this index
satisfies the minimum length.
The process is repeated to obtain batch_size trajectories. The tensors
of the individual trajectories are then stacked together to create the
training batch.

Also, stacking the data tensors in the batch dimension is now outsourced
to TrainingData.batch_stack(). This makes the replay buffer code more
concise since it concentrates on the core logic.

Performance improvements:

Previously, a list of trajectories was stored, with each one potentially
having a different length. To sample a trajectory subsection, first a
trajectory was selected, and then a start index inside it.
Drawing a random trajectory was weighted by their lengths to account for
longer trajectories containing more states / information. This used
numpys random.choice(p=...) which I benchmarked to be a very costly
operation. Also, all lengths had to be read into an array, incurring a
O(n) cost.
In comparison, pythons random.randrange() is orders of magnitude faster,
especially when the numbers (buffer length) become big. By drawing a
random position inside the contiguous buffer, longer trajectories
automatically have a higher chance of being selected.

Furthermore, the new ring buffer has only O(1) random access cost. This
is much better than the previously used deque, which is implemented as a
linked list, providing random access in O(n).
The performance was further crippled by reading the whole deque into a
numpy array.