Remove early exit check in training unroll loop
At the time this was written, the batch was padded with dummy trajectory
states up to the the configured length. To not waste compute, the check
exited the training unroll loop early when all trajectories ended.

The replay buffer generating the training batch was rewritten. The new
implementation truncates the batch (trajectory) length automatically
when all trajectories ended, making the check superfluous.