Unroll with dynamics network latent during training
Train the networks like they are used during the tree search:
The latents are recurringly predicted using the dynamics network since
no intermediate observations are available.
This follows the original MuZero implementation.

The dynamics network now receives gradients through the prediction
network, which was not the case in single player games with observations
every step.
This improves the predictions during the tree search.
It also enables training with a latent loss weight of zero.
In fact, the training of the dynamics network only relied on gradients
from the latent loss previously.

I swapped the if/else branch because:
- I like the positive logic in the if branch more
- the if branch is shorter, so one can see more quickly the condition
  for the else branch