Speed up loss average and sum
This should yield a speed improvement for big tensors:
Performing the mean() first reduces the tensor size to the batch size,
so the sum() is calculated over less elements.

The view() is used over flatten(start_dim=1) because the loss tensor
might only have a single dimension so the call would fail.