Implement __iter__() for RingBuffer
I just wrote it because I thought need it during debugging, now I don't
want to throw the work away.
It is not used in the code yet.
Iterating worked previously as well because python falls back to the
sequence protocol using __len__() and __getitem__(), but this should be
faster since the list is implemented in CPython.