Force intended tensor dtypes in TrainingData
This is more for good measure, since torch probably does the right thing
even with the wrong or mixed dtypes.
Although the code has type annotations and passes mypy, this can't be
relied on:
I happend to have a case where hydra interpreted a config item as int,
passed it to the game class during instantation. This int wound up to be
used for the return value of the reward, which eventually created
tensors with dtype int.
Since the value was only created at runtime by hydra and passed on into
the class __init__ with the wrong type, mypy was not able to catch that.

See the next commit for details.