Ensure argument float types in OpenSpielGame at runtime
This prevents a very subtle type discrepancy which was not catched by the
type checker:
The arguments to OpenSpielGame are usually provided through hydra when it
instantiates the class with parameters from the configuration.
In my case, the problematic config item was:
bad_move_reward: -1
Since hydra infers the data type of config items from its value, it
passed an int value to the class, which then got stored.
It was returned as the reward for an invalid move, got passed around and
ended up in a torch.tensor() call. Torch inferred this should become an
int tensor.
Not problematic, but also not nice.

See also the previous commit for a fix in TrainingData.