Fix bug in policy fn FromExpandedValues: Add reward
The immediate value of a node is its reward + value.
The value predictions alone are only a measure of the expected future
rewards beyond this node / state.

Funnily, the algorithm learned previously with this bug.
Now I fixed a couple of other bugs (see previous commits) and
implemented terminal nodes (see later commits), which exposed this
pretty substantial bug.