Replace Self type annotations with concrete types
I recently learned that Self is a dynamic TypeVar with an upper bound of
the current class.
The difference to the concrete class types are relevant for subclassing.

Since nothing in that file is subclassed, I changed it to the concrete
types to help readability in mypy error messages:
"Self" has no attribute "..."