Fix value target calculation to MuZero version
MuZero calculates a different value target than classic RL, which is
where I looked up the formulas last time.

This updates the code and comments to reflect the correct behavior.

This time I am confident that the calculations are correct, since they
are sound in itself (like extending a nstep horizon with the value
target) and equivalent to what the tree search does.