Improve performance in TrainingData.stack_batch()
I discovered that using attrgetter is up to 4x faster than using
attrs.astuple.
I expect around 4us gained per instance.

The change also makes clear that the fields are extracted in the same
order for all instances, which is crucial to stacking the data.
It worked previously, but the attrs docs are not clear if the order is
guaranteed.