Move loss averaging to dunder method in LossCounts
This has two benefits:
- visually declutters the main training code
- agnostic to the number and naming of loss and count terms, enabling
  easier future changes

It comes at two drawbacks:
- code is harder to read
- implementation not easily typecheckable
These drawbacks should be acceptable since the method has some
documentation and is relatively small.

This is implementated as a rtruediv as opposed to a truediv on Losses
because it is an implementation detail of LossCounts:
Losses does not really need to know that there are counts by which it
can be divided.
Put differently, if LossCounts were to be removed, nothing has to change
in Losses.