Improve signatures of networks to typecheck their usage
This change provides the correct function signatures for __call__() and
si() in the network classes. This enables typechecking the uses /
calling locations of the networks in the rest of the code.

It immediately uncovered a bug in PerfectInformationRLPlayer.own_move(),
where the representation network was called with an additional None
argument before the observations. This is a remnant from the time with
beliefs in the representation network. It did not do any harm because
None inputs are filtered out in the fully connected networks.

Previously:
- forward() had the correct signature, but that is never called in the
  code, only overridden by network implementations
- __call__()'s signature came from nn.Module, where it has Any
  for arguments and return value
- si() came from NetworkBase, where I could only customize the return
  type

An ideal solution would be to have base class which is generic over
function signatures, so the code can be reused.
I was not able to get that to work in mypy tho:
https://github.com/python/mypy/issues/15096

So, therefore a lot of duplicate code in each of the abstract base
classes.