Fix mcts backpropagation: Include omitted leaf node
The backpropagation was "off-by-one" in the sense that new leaf nodes
did not get touched at all:
A newly-created leaf node had a visit count of zero and also no value
samples. It's value should be the prediction for its state, which also
makes more sense.

This has always been MuZero's original implementation, but their
convoluted pseudocode made it hard to see what is going on.