Move most mcts node attributes into base class
Move all attributes into the base class which are derived from the
dynamics network.
Also rename the classes so that Node now refers to the
abstract base class. This way 'Node' instances are still guaranteed to
have all relevant attributes, so most usage locations do not need to be
updated despite the refactor.

The previous NodeBase class was effectively removed. An abstraction this
general is not needed anymore since at the moment there are no other
node types than the ones mapping to game states.

The dict was replaced with a MutableMapping because the latter is
covariant in its type of arguments, which is required when assigning
subclasses of Node to the children.