Replace Optional beliefs with valid tensors
Instead of using Optional[Tensor] for the beliefs, always use a valid
tensor or tensor shape. To disable beliefs, insert a zero data dimension
in the belief shape.

This has the benefit of uniform handling of beliefs next to the latent,
with no specialities like None Tensors or shapes.
It enables to trace the network implemenation to TorchScript. Traces may
not contain None inputs or outputs, but zero dimension tensors are okay.

Using zero-dim beliefs was the original approach, but I switched to
optional Tensors because displaying a model graph in the tensorboard
proved difficult for zero-dim tensors.
In the mean time, I found out that this is a tensorboard bug I can work
around much easier than trying to jit/trace models with None in/outputs.