Refactor fc nets implementation, fix return shapes
The networks now use the in/output shapes as defined by the containers,
with a unified implementation to split the output tensors.
I also fixed a bug where the returned tensors only have 1D shapes.
That is, multiple dimensions for the latent or belief were not respected
and only a flattened tensor was returned.

This problem did not reveal itself because I only ever tested with 1D
latent shapes.