Refactor net containers: Split train/inference forward
The key realisation was that I can use multiple methods instead of
cramming everything into the main forward().
Specifically, there are now three methods on each container:
- forward(): normal forward for inference of batched data
- si(): single inference for unbatched data
- raw_forward(): directly calls the wrapped network implementation,
  useful for training since the raw logits are returned

The si() method allows for the removal of the si submodule.
The raw_forward() allows to remove the logits=true kwargs.
Together, this allows for a great refactor and simplifications of the
network container and the jit logic.
Everything can be traced now too, no more scripting necessary.

The refactor is good because it gets rid of nested si containers and
autobatch modules.
The si containers were nasty because they were special instances of the
containers. They had no si attribute and required special handling in
jit() and __init__().