Implement value and reward Rescaler
To make the networks compatible with reward and values of any range, a
Rescaler Module is introduced.
It translates between the [-1, 1] range (the network outputs) to the
actual range. The actual range is determined by the minimum and maximum
of all samples currently in the replay buffer.

Directly predicting values of large magnitude leads to exploding
gradients during training and is avoided with the Rescaler.