Remove in/output shapes from network containers
Exposing the network in/output shapes only proved beneficial for a fully
connected network, where in/output sizes can be determined
programatically.
All other networks architectures very likely need to handle the various
in/outputs differently or individually, so in code it is clearer to get
the shape information from the respective config items directly.
This increases code duplication, but should improve code readability.

To reduce the coupling between the config and the containers / fully
connected networks, the config is now only imported when the shape
information is accessed.