Move abbrev defs to yaml + use exclamation marks keys
This commit introduces two major changes to the abbreviation /
substitution system:

First, the definitions of the substitutions are moved into a yaml file.
This provides a better overview and allows for later extensions of the
abbreviation data, like adding explanations for a glossary.
It also allows to check if keys are clashing.

Second, the abbreviations in the text must be preceeded by an
exclamation mark, like !key.
This provides multiple benefits:
- makes it explicit where an expansion will happen
- typos are less likely to go unnoticed since invalid keys can be
  detected as opposed to just plain text
- stuff is not expanded on accident, like in the bibliography listing
- the previous point enables to use shorter keys, like !a for 'action'

Some new substitutions are also added, and some others have their keys
shortened.
The text changes are best viewed with a word diff.

An implementation detail is that there is now only one show rule
performing the substitutions, instead of multiple. This should improve
the performance because previously, the document was processed multiple
times, once for every substitution.

Also, the pre-commit config is adjusted to not match any yaml files
except the bibliography.