Remove beliefs
I think beliefs were meant as a way to transport latent information
between states in a way that is not coupled to the latent representation
of observations.
However, this might not matter anymore because the scope of this project
is reduced to games of perfect information.
These already provide all information in observations, so beliefs are
not necessary.

In practice, I never came to experiment with beliefs since I already
struggled with getting the basic learning to work.
Also, the implementation with belief shapes are annoying because in a
multidimensional case the dimensions of latent and belief must match.

Classic example of YAGNI.

Sorry for all the confusion in the git history with belief tensors all
over the place.