Move recursive jit() logic into Networks class
Incidentally, this also fixes the tensorboard graph for the Rescaler:
Now, it is displayed as a submodule with all the computations happening
inside it and with correct in/output shapes.
Previously, the Rescaler already was jitted as a standalone module when
the container itself was traced, which confused the tensorboard graph
algorithm.