Implement SLAW (automatic loss weighting)
It is disabled in the config because I found it to perform worse than without
(at least in my test cases).

If this commit seems a bit useless without enabling SLAW, it might be because I
found out the performance regression months later.
I amended this commit (disabling SLAW in the config) and left the code in-place
to avoid merge conflicts during the rebase.