Remove initial latents / Train only on observations
Since the scope of the project is definitely only perfect information
games, storing intermediate latents to not make much sense anymore.
The agent restarts its calculations anew on each observation.

They are cumbersome at best and hurt the training at worst, because in a
multiplayer setting, old latents are stored in the replay buffer and get
used during training.