Predict per-player rewards & values for max-n search
The original MuZero algorithm is limited to single-player games and
two-player zero-sum games.
The zero-sum assumption is build deep into the architecture, because the
tree search is effectively negamax (minimax for two players).
The "correct" search behaviour in the game tree is the max-n search
which maximizes the payoff for each player individually.

Max-n in MuZero consequently requires reward and value network
predictions for every player individually.
This commit updates the application to predict and train reward and
value vectors instead of scalars.

This change also removes the reward fn, the reward vector is taken
directly from the game implementation.