Fix small rescaler ranges giving NaN
In unfortunate cases, the replay buffer may be filled with games that
all have the same reward / value.
This gives max = min, so when normalizing values by dividing with (max -
min), this leads to NaN.
To avoid that the rescaler bounds are too close together, add a small
number for numerical stability.