Fix citation brackets
The previous solution did not add square brackets around multiple
citations at once.
This is because cite() is directly called with multiple keys, and the
ref show rule does not cover it.

The new solution consists of a show rule that add square brackets around
all cite() calls when the brackets parameter is set.
These outer brackets are then enabled with a set rule.
citet() overrides the brackets to false and adds its own around the year
number.

The nonbreaking space is removed since the citations with author and
year act as words on their own (unlike only numerical keys).