Improve map path handling
If only a map name is given, the map is attempted to be loaded from the
default path (the maps/ directory).
The maps/ directory is derived from the path of the map.py file, and not
relative to the working directory anymore.
An absolute path can also given to load maps from somewhere else.